Pimp out Candidate List data
Incorporated new candidate data. Candidates with facebook and/or twitter
urls will have buttons to profiles added. "Info" buttons will now point
toward candidate website, if it's in the sheet, or OR if it's not.

Additionally, candidate images now have the same behavior as website,
pulling from the csv or defaulting to OR